Extract opentaint-ir-storage module (#271)
[opentaint-ir-storage] Extract opentaint-ir-api-storage module

The opentaint-ir-api-storage module depends on opentaint-ir-api-common.
The opentaint-ir-api-jvm module depends on opentaint-ir-api-common & opentaint-ir-api-storage.

[opentaint-ir-storage] Fix review points: remove EntityRelationshipStorage.asReadWrite

[opentaint-ir-storage] Fix review points: setImmutable() waits for background jobs

[opentaint-ir-storage] Minor: remove some comments

[opentaint-ir-storage] Minor: rename

[opentaint-ir-api, opentaint-ir-core] Add ability to set persistence and database to immutable state

Got rid of the JIRDatabase.classpathOf() method.

[opentaint-ir-storage] Add navigation over AttributesImmutable

In implementation of AttributesImmutable, AttributesCursor added for navigation
in attribute values. It is used in PropertiesImmutable for searching by property
value and for searching in a range of values. AttributesCursorEntityIterable added
as a lazy implementation of EntityIterable atop of AttributesCursor.

More test cases added.

[opentaint-ir-storage] Generate and add SparseBitSetTest

[opentaint-ir-core] Extract ERS API implementations to module opentaint-ir-storage

All EntityRelationshipStorage implementations, except SqlEntityRelationshipStorageSPI,
moved to the opentaint-ir-storage module. Extra minor refactoring are done in order to
de-couple dependencies.

[opentaint-ir-core] Get rid of unnecessary extra check if instanceIds are sorted

[opentaint-ir-core] Add first version of RAMDataContainerImmutable

RAMDataContainerImmutable is intended as a data container for RAMEntityRelationshipStorage
saving data in a compact format with high data locality using large arrays.

Auxiliary refactorings:
- EntityIdSet is obsolete;
- EntityIdCollectionEntityIterable is obsolete,
as it is replaced with InstanceIdCollectionEntityIterable everywhere;
- SparseBitSet added for compact bitsets;
- ByteArray.asComparable() added;
- ERS API: Transaction.getEntityUnsafe(id) added.

TODO: provide PropertiesImmutable with property value indices and use then in
find*() methods. Currently, find*() methods use brute force search.

[build] Add facilities to deal with read-only state of storages

For EntityRelationshipStorage, added ability to get read-only & read/write
snapshots of the storage. By default, they do nothing.

In PluggableKeyValueStorage, added property allowing to switch to read-only
mode and back to read/write mode. It has non-trivial implementation only in
XodusKeyValueStorage.

[build] For tests, set JVM target to that the build is being run under

This change allows to run tests against examples compiled for different JVM
targets whereas main artifacts are still being compiled for Java 1.8.
Both Java and Kotlin configured. We have to compile test fixtures for 1.8
as well since Gradle considers them as dependencies of main artifacts.

[opentaint-ir-core] Update Kotlin to version 2.0.20

As of this commit, Kotlin metadata version is equal to Kotlin version.
Refactored Kotlin metadata classes' usages as they live in the new package.

Fixed nullability issues with Kotlin class types appeared due to lack
of corresponding JB annotations since Kotlin 2.0.0. Extension functions
declared in the Nullables.kt moved from the opentaint-ir-api-jvm module to
the opentaint-ir-core one.

[opentaint-ir-core] Fix running benchmarks with ERS persistence using pluggable K/V storages

[opentaint-ir-core] Process Class-Path attribute of jar manifest

As of this commit, classpath is not only set of jar files and build directories,
but also a set of references to jar files calculated using "Class-Path" attribute
in manifest of any jar file.